chat.tsxとhome.tsxのトークンチェックの外だし
