+++ pages/api/home/home.tsx
@@ -43,7 +43,7 @@ import { HomeInitialState, initialState } from './home.state';
 import { jwtDecode } from "jwt-decode";
 
 import { v4 as uuidv4 } from 'uuid';
-
+import { isTokenExpired, refreshJWTbytoken } from '../auth/token/tokencheck';
 
 interface Props {
   serverSideApiKeyIsSet: boolean;
@@ -226,26 +226,17 @@ const Home = ({
   //   return newToken;
   // };
   
-  // jwtを外部キャッシュでリフレッシュするAPIリクエスト
-  const refreshJWTbytoken = async (oid: string) => {
-    const url = "/api/auth/verify";
-    const { data }: { data: AuthenticationResult } = await axios.put(url, {
-      oid
-    });
-    const newToken = data.accessToken;
-    setJWT(newToken);
-    return newToken;
-  };
 
-// トークンの有効期限をチェックする関数
-const isTokenExpired = (token: string) => {
-  const payload = JSON.parse(Buffer.from(token.split('.')[1], 'base64').toString('utf-8'));
-  // console.log("expiration:",payload.exp * 1000," now:",Date.now())
-  const issuedAt = payload.iat * 1000; // iatは秒単位なのでミリ秒に変換
-  const ninetySeconds = 9000 * 1000; // 90秒をミリ秒に変換
-  return Date.now() > (issuedAt + ninetySeconds);
-  // return payload.exp * 1000 < Date.now();
-};
+
+// // トークンの有効期限をチェックする関数
+// const isTokenExpired = (token: string) => {
+//   const payload = JSON.parse(Buffer.from(token.split('.')[1], 'base64').toString('utf-8'));
+//   // console.log("expiration:",payload.exp * 1000," now:",Date.now())
+//   const issuedAt = payload.iat * 1000; // iatは秒単位なのでミリ秒に変換
+//   const ninetySeconds = 9000 * 1000; // 90秒をミリ秒に変換
+//   return Date.now() > (issuedAt + ninetySeconds);
+//   // return payload.exp * 1000 < Date.now();
+// };
 
   const fetchModels = useCallback(async (signal?: AbortSignal) => {
     let token = jwt;
@@ -266,6 +257,7 @@ const isTokenExpired = (token: string) => {
         console.log("put_start")
         // const jsonrefreshtoken: string = JSON.parse(storedrefreshtoken);
         token = await refreshJWTbytoken(storedoid);
+        setJWT(token);
         console.log("put_end", token)
       }
     }
